// vendor/codexapp-frontier-src/src/types/codex.ts
export type RpcEnvelope<T> = {
  result: T
}
// ...
export type RpcMethodCatalog = {
  data: string[]
}